refactor(llm): extract litellm_client types + stateless leaves (Tier-2.5 Task 1)
Move to three dependency-free sibling leaf modules, re-exported by import
binding (SINK-2, identity-preserving) so the facade surface is unchanged:

- _litellm_types.py: LiteLLMConfig, ToolCallingChatResponse, LiteLLMClientError,
  StructuredOutputParseError, LLMHardTimeoutError.
- _litellm_subprocess.py: the 6 _Completion*Snapshot dataclasses + snapshot
  builders + _litellm_completion_worker (calls litellm.completion via the module
  attr so the fork-inherited global LLM mock still intercepts).
- _litellm_json_extraction.py: the 6 pure JSON helpers (were self-free methods;
  now module-level free functions) + _PYTHON_TO_JSON_REPLACEMENTS.

The facade re-imports every moved name (the 6 snapshots by redundant-alias
re-export) and _maybe_parse_structured_output now calls the JSON free functions.
Surface guard extended: snapshot-class identity asserts now enforce each facade
_Completion*Snapshot IS the _litellm_subprocess class. Zero behavior change,
zero SINK-1 (no patch-namespace move). test_litellm_client_unit updated to call
the two directly-tested JSON helpers as free functions (imported from the facade,
same objects).
